+++ src/model/output/decoded/token.rs
@@ -8,6 +8,10 @@ use crate::model::pipeline::context::EntityContext;
 use crate::model::output::tensors::TensorOutput;
 use super::SpanOutput;
 
+
+const TENSOR_LOGITS: &str = "logits";
+
+
 /// Decoding method for token mode.
 /// 
 /// From the related (GLiNER multi-task) [paper](https://arxiv.org/pdf/2406.12925v1): 
@@ -29,13 +33,17 @@ impl TensorsToDecoded {
         }
     }
 
+    pub fn outputs() -> [&'static str; 1] {
+        [TENSOR_LOGITS]
+    }
+
     fn decode(&self, input: &TensorOutput) -> Result<Vec<Vec<Span>>> {
         // prepare output vector
         let batch_size = input.context.texts.len();
         let mut result: Vec<Vec<Span>> = std::iter::repeat_with(Vec::new).take(batch_size).collect();
 
         // look for logits and check its shape
-        let logits = input.tensors.get("logits").ok_or("logits not found in model output")?;
+        let logits = input.tensors.get(TENSOR_LOGITS).ok_or("logits not found in model output")?;
         self.check_shape(logits.shape()?, &input.context)?;
     
         // extract the actual array