+++ src/model/pipeline/span.rs
@@ -1,5 +1,6 @@
 //! Pre-defined pipeline for NER (span mode)
 
+use std::collections::HashSet;
 use std::path::Path;
 use ::composable::*;
 use orp::{pipeline::*, params::RuntimeParameters};
@@ -13,6 +14,8 @@ use super::context::EntityContext;
 pub struct SpanPipeline<S, T> {
     splitter: S,
     tokenizer: T,
+    expected_inputs: HashSet<&'static str>,
+    expected_outputs: HashSet<&'static str>,
 }
 
 impl<'a, S: Splitter, T:Tokenizer> Pipeline<'a> for SpanPipeline<S, T> {
@@ -39,6 +42,15 @@ impl<'a, S: Splitter, T:Tokenizer> Pipeline<'a> for SpanPipeline<S, T> {
             output::decoded::greedy::GreedySearch::new(params.flat_ner, params.dup_label, params.multi_label)
         ]
     }
+
+    fn expected_inputs(&self) -> Option<&std::collections::HashSet<&str>> {
+        Some(&self.expected_inputs)
+    }
+
+    fn expected_outputs(&self) -> Option<&std::collections::HashSet<&str>> {
+        Some(&self.expected_outputs)
+    }
+
 }
 
 /// Specific implementation using HF tokenizer and default splitter
@@ -47,6 +59,8 @@ impl SpanPipeline<crate::text::splitter::RegexSplitter, crate::text::tokenizer::
         Ok(Self {
             splitter: crate::text::splitter::RegexSplitter::default(),
             tokenizer: crate::text::tokenizer::HFTokenizer::from_file(tokenizer_path)?,
+            expected_inputs: crate::model::input::tensors::span::SpanTensors::inputs().into_iter().collect(),
+            expected_outputs: crate::model::output::decoded::span::TensorsToDecoded::outputs().into_iter().collect(),
         })
     }
 }